+++ src/Firebase/AppCheck/AppCheckTokenGenerator.php
@@ -8,10 +8,11 @@ use Beste\Clock\SystemClock;
 use Firebase\JWT\JWT;
 use Kreait\Firebase\Exception\AppCheck\InvalidAppCheckTokenOptions;
 use Psr\Clock\ClockInterface;
-use SensitiveParameter;
 
 /**
  * @internal
+ *
+ * @todo Add #[SensitiveParameter] attribute to the private key once the minimum required PHP version is >=8.2
  */
 final class AppCheckTokenGenerator
 {
@@ -25,7 +26,7 @@ final class AppCheckTokenGenerator
      */
     public function __construct(
         private readonly string $clientEmail,
-        #[SensitiveParameter] private readonly string $privateKey,
+        private readonly string $privateKey,
         ?ClockInterface $clock = null,
     ) {
         $this->clock = $clock ?? SystemClock::create();